Migrated UpdateUserPasswordInDatabaseUsersTableAndRespondJsonOfUserHandler() to a new file update_user_password_handler.
